Move `BlockTree` and `PointSchedule` out of `ChainGenerator.Tests`
