+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -518,13 +518,12 @@ test-suite infra-test
     Ouroboros.Consensus.Util.Tests
     Ouroboros.Network.AnchoredFragment.Extras
     Ouroboros.Network.Driver.Limits.Extras
+    Test.Ouroboros.Consensus.BlockTree
     Test.Ouroboros.Consensus.ChainGenerator.Tests
     Test.Ouroboros.Consensus.ChainGenerator.Tests.Adversarial
     Test.Ouroboros.Consensus.ChainGenerator.Tests.BitVector
-    Test.Ouroboros.Consensus.ChainGenerator.Tests.BlockTree
     Test.Ouroboros.Consensus.ChainGenerator.Tests.Counting
     Test.Ouroboros.Consensus.ChainGenerator.Tests.Honest
-    Test.Ouroboros.Consensus.ChainGenerator.Tests.PointSchedule
     Test.Ouroboros.Consensus.Genesis.Setup
     Test.Ouroboros.Consensus.Genesis.Tests
     Test.Ouroboros.Consensus.Genesis.Tests.LongRangeAttack
@@ -535,6 +534,7 @@ test-suite infra-test
     Test.Ouroboros.Consensus.PeerSimulator.Run
     Test.Ouroboros.Consensus.PeerSimulator.ScheduledChainSyncServer
     Test.Ouroboros.Consensus.PeerSimulator.Trace
+    Test.Ouroboros.Consensus.PointSchedule
     Test.Util.ChainUpdates.Tests
     Test.Util.Schedule.Tests
     Test.Util.Split.Tests
